requirements.txt
Flask==3.0.3
Pillow==12.0.0
pycryptodome==3.20.0
gunicorn==22.0.0
